+++ Service/Vault/GetSavedCards.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 namespace Yireo\LokiCheckoutMollie\Service\Vault;
 
 use Magento\Customer\Model\Session;
-use Magento\Framework\Api\SearchCriteriaBuilderFactory;
+use Magento\Framework\Api\SearchCriteriaBuilder;
 use Magento\Framework\Serialize\SerializerInterface;
 use Magento\Vault\Api\Data\PaymentTokenInterface;
 use Magento\Vault\Api\PaymentTokenRepositoryInterface;
@@ -12,10 +12,10 @@ use Magento\Vault\Api\PaymentTokenRepositoryInterface;
 class GetSavedCards
 {
     public function __construct(
-        private SearchCriteriaBuilderFactory    $searchCriteriaBuilderFactory,
+        private SearchCriteriaBuilder $searchCriteriaBuilder,
         private PaymentTokenRepositoryInterface $paymentTokenRepository,
-        private SerializerInterface             $serializer,
-        private Session                         $customerSession
+        private SerializerInterface $serializer,
+        private Session $customerSession
     ) {
     }
 
@@ -35,14 +35,13 @@ class GetSavedCards
             return [];
         }
 
-        $search = $this->searchCriteriaBuilderFactory->create();
-        $search->addFilter(PaymentTokenInterface::IS_VISIBLE, 1);
-        $search->addFilter(PaymentTokenInterface::IS_ACTIVE, 1);
-        $search->addFilter(PaymentTokenInterface::CUSTOMER_ID, $this->customerSession->getCustomerId());
-        $search->addFilter(PaymentTokenInterface::PAYMENT_METHOD_CODE, 'mollie_methods_creditcard');
+        $this->searchCriteriaBuilder->addFilter(PaymentTokenInterface::IS_VISIBLE, 1);
+        $this->searchCriteriaBuilder->addFilter(PaymentTokenInterface::IS_ACTIVE, 1);
+        $this->searchCriteriaBuilder->addFilter(PaymentTokenInterface::CUSTOMER_ID, $this->customerSession->getCustomerId());
+        $this->searchCriteriaBuilder->addFilter(PaymentTokenInterface::PAYMENT_METHOD_CODE, 'mollie_methods_creditcard');
 
         $output = [];
-        $items = $this->paymentTokenRepository->getList($search->create())->getItems();
+        $items = $this->paymentTokenRepository->getList($this->searchCriteriaBuilder->create())->getItems();
         foreach ($items as $item) {
             $details = $this->serializer->unserialize($item->getTokenDetails());
             $details['public_hash'] = $item->getPublicHash();